Got rid of renaming crap, passed all names through in `subtest._message`. Bugfixes to `NonRadicalTotalUtilityAxiom`. `TestVotingAlgorithmAgainstAxioms` runs and logs. Getting some `invalid value encountered in scalar divide` warnings in `utility_tradeoff` nested function in `NonRadicalTotalUtilityAxiom`. Added numerical checks to avoid divide by zero, so not sure what that is.

+++ generative_social_choice/slates/voting_utils.py
@@ -119,7 +119,7 @@ def pareto_efficient_slates(
     for slate in metric_values.index:
         utilities = voter_max_utilities_from_slate(rated_votes, slate)
         for metric_index, metric in enumerate(positive_metrics):
-            metric_values.at[slate, metric_index] = metric(utilities).astype(np.float64)
+            metric_values.at[slate, metric_index] = metric(utilities)
     
     return set(frozenset(cand_tuple) for cand_tuple in metric_values.index[is_pareto_efficient(metric_values.values)])
